+++ src/components/header/Header.tsx
@@ -6,10 +6,10 @@ import img_delete from "../../img/ic_delete.png";
 import img_help from "../../img/ic_help.png";
 import img_search from "../../img/ic_search.png";
 import img_setting from "../../img/ic_settings.png";
+import { logger } from "../../logger";
 import { createEvent, createIdleEvent } from "../../model/event";
 import * as action from "../../redux/actions";
 import { selectMapState, selectStationState } from "../../redux/selector";
-import { logger } from "../../script/logger";
 import "./Header.css";
 import StationSearchBox, { StationSuggestion } from "./StationSearchBox";
