+++ src/components/map/Map.tsx
@@ -5,9 +5,9 @@ import { CSSTransition } from "react-transition-group"
 import pin_location from "../../img/map_pin.svg"
 import pin_station from "../../img/map_pin_station.svg"
 import pin_station_extra from "../../img/map_pin_station_extra.svg"
+import { logger } from "../../logger"
 import { selectMapState, selectStationState } from "../../redux/selector"
 import StationService from "../../script/StationService"
-import { logger } from "../../script/logger"
 import { CurrentPosDialog } from "../dialog/CurrentPosDialog"
 import { LineDialog } from "../dialog/LineDialog"
 import { StationDialog } from "../dialog/StationDialog"
