+++ src/components/map/mapEventHook.ts
@@ -1,6 +1,7 @@
 import { MutableRefObject, useRef, useState } from "react"
 import { useDispatch, useSelector } from "react-redux"
 import { useSearchParams } from "react-router-dom"
+import { logger } from "../../logger"
 import { parseQueryBoolean } from "../../model/diagram"
 import { Line } from "../../model/line"
 import { LatLng } from "../../model/location"
@@ -9,7 +10,6 @@ import * as action from "../../redux/actions"
 import { selectMapState } from "../../redux/selector"
 import { AppDispatch } from "../../redux/store"
 import StationService from "../../script/StationService"
-import { logger } from "../../script/logger"
 import { useRefCallback } from "../hooks"
 import { NavType, isStationDialog } from "../navState"
