+++ src/components/map/mapHook.ts
@@ -1,5 +1,6 @@
 import { MutableRefObject, RefObject, useEffect, useState } from "react"
 import { useDispatch, useSelector } from "react-redux"
+import { logger } from "../../logger"
 import { PolylineProps, RectBounds, getBounds, getZoomProperty, isInsideRect } from "../../model/diagram"
 import { Line } from "../../model/line"
 import { LatLng, MapCenter } from "../../model/location"
@@ -8,7 +9,6 @@ import * as action from "../../redux/actions"
 import { selectMapState } from "../../redux/selector"
 import { AppDispatch } from "../../redux/store"
 import StationService, { DataType } from "../../script/StationService"
-import { logger } from "../../script/logger"
 import { useRefCallback } from "../hooks"
 import { NavType, isStationDialog } from "../navState"
 import { useHighVoronoi } from "./voronoiHook"
