--- src/script/type.d.ts
@@ -1,7 +1,0 @@
-declare module "worker-loader!*" {
-  class VoronoiWorker extends Worker {
-    constructor();
-  }
-
-  export default VoronoiWorker;
-}
