+++ src/test/unit/kdTree.test.ts
@@ -1,7 +1,7 @@
 import { LatLng } from "../../model/location"
 import { Station, StationAPIResponse, parseStation } from "../../model/station"
 import { StationTreeSegmentResponse, isStationLeafNode } from "../../script/StationService"
-import { NodeProps, StationKdTree, StationLeafNodeProps, StationNode, StationTreeSegmentProps } from "../../script/kdTree"
+import { NodeProps, StationKdTree, StationLeafNodeProps, StationNode, StationTreeSegmentProps } from "../../search/kdTree"
 
 describe("kdTree", () => {
   let data: StationTreeSegmentResponse
