+++ src/worker/type.d.ts
@@ -1,0 +1,7 @@
+declare module "worker-loader!*" {
+  class VoronoiWorker extends Worker {
+    constructor();
+  }
+
+  export default VoronoiWorker;
+}
